nambah peminjaman

+++ routes/web.php
@@ -22,9 +22,13 @@ Route::get('logout', function() {
 
 Route::get('/home', 'HomeController@index')->name('home');
 
-Route::get('/inventaris', 'InventarisController@index')->name('inventaris.index');
-Route::get('/inventaris/create', 'InventarisController@create')->name('inventaris.create');
-Route::post('/inventaris/store', 'InventarisController@store')->name('inventaris.store');
-Route::get('/inventaris/{id}/edit', 'InventarisController@edit')->name('inventaris.edit');
-Route::put('/inventaris/{id}/update', 'InventarisController@update')->name('inventaris.update');
-Route::delete('/inventaris/{id}/delete', 'InventarisController@delete')->name('inventaris.delete');
+Route::get('/inventaris', 'Admin\InventarisController@index')->name('inventaris.index');
+Route::get('/inventaris/create', 'Admin\InventarisController@create')->name('inventaris.create');
+Route::post('/inventaris/store', 'Admin\InventarisController@store')->name('inventaris.store');
+Route::get('/inventaris/{id}/edit', 'Admin\InventarisController@edit')->name('inventaris.edit');
+Route::put('/inventaris/{id}/update', 'Admin\InventarisController@update')->name('inventaris.update');
+Route::delete('/inventaris/{id}/delete', 'Admin\InventarisController@delete')->name('inventaris.delete');
+
+Route::get('/peminjaman', 'User\PeminjamanController@index')->name('peminjaman.index');
+Route::get('/peminjaman/create', 'User\PeminjamanController@create')->name('peminjaman.create');
+Route::post('/peminjaman/store', 'User\PeminjamanController@store')->name('peminjaman.store');
